build based on 653a2f8

+++ previews/PR10/literal/example.jl
@@ -76,8 +76,8 @@ fig = Figure()
 z1 = [significant_digits(f(precify(xi), precify(yi))) for yi in y, xi in x]
 z2 = [significant_digits(f_improved(precify(xi), precify(yi))) for yi in y, xi in x]
 
-zmin = floor(Int64, min(minimum(z1), minimum(z2)))
-zmax = ceil(Int64, max(maximum(z1), maximum(z2)))
+zmin = floor(Int, min(minimum(z1), minimum(z2)))
+zmax = ceil(Int, max(maximum(z1), maximum(z2)))
 
 ax1 = Axis(fig[1, 1]; aspect = AxisAspect(1), title = "Original f", xticksvisible = false, yticksvisible = false, xticklabelsvisible = false, yticklabelsvisible = false)
 ax2 = Axis(fig[1, 2]; aspect = AxisAspect(1), title = "Improved f", xticksvisible = false, yticksvisible = false, xticklabelsvisible = false, yticklabelsvisible = false)
